requirements.txt
numpy
Pillow
pycryptodome